refactor: content script async/await and blocking embeds
`timer.js` is now `bouncer.js`, and the content script makes use of
async/await when using extension APIs.

The bouncer content script now works on `all_frames`, so embedded sites
are subject to blocked rules. This may result in a performance penalty,
since currently the bouncer script is injected into frames with any URL.
It may be necessary to use a background script to create dynamic content
script rules based on the site blocking rules.