feat!: first functional policies
First version of Bouncer with policies that function (though the
implementation is far from ideal).

To try it, after giving the extension full permissions, go to its
settings page and add a rule with a valid host (milliseconds don't
matter). Visit a page with that host, and it should be blocked by
default.