feat!: POC for viewtime limit with cooldown
Adds a new limit type: limits to an amount of viewtime, after which the
site is blocked for the specified cooldown period. The limit is
untested, and is mainly a proof of concept (though the code functions
well).

Playing with this limit reveals a serious bug that I feared was present.
To reproduce:

1. Place a viewtime limit on a given host (www.example.com)
2. Open two tabs visiting the limited site
3. Before a block is applied to either tab, switch directly between the
   tabs.

Expected behavior: the viewtime on both tabs accumulates, so you are
limited to the combined time spent on both tabs.

Actual behavior: switching between tabs does not cause the time to
accumulate, and up to double the allowed viewtime is permitted.

This bug is caused by a faulty assumption: I assumed that during a tab
switch, the pagehide handler of the tab being hidden would complete
before the content script started in the newly opened tab. Because this
isn't the case, the old tab may not have stored its new total viewtime
before the new tab reads the page data. This will have to be fixed in a
future commit.