feat!: working blocker using background script
This version of Bouncer uses a background script as the "server" for
content script "clients". The content script send messages to the
background script, which reads/writes Bouncer data and sends back
commands.

This version lays the groundwork for handling bugs around concurrent
view sessions on the same page object, but doesn't yet fix those bugs.
There also needs to be more documentation and typing around the
messaging between content and background scripts.