fix: sender id improvements
This commit fixes issues with sender IDs used to populate page viewers. The window ID is no longer used in constructing sender IDs, as it can change independently. The background and content scripts also handle the case where the sender ID fields are unpopulated when a tab or window is closed.