refactor: improved data context abstraction
This commit introduces new types and classes for handling Bouncer's
persisted data.

The IContext and PersistedContext expose runtime data to the running
extension.

The IDataMapper and StorageDataMapper is a translation layer that hides
persistent storage and data from consumers of the runtime objects that
data represents.

Future versions of the code will build on these abstractions to handle
the potential for runtime data loss when the background service worker
idles. Currently, the background controller "forgets" the guards
assigned to each tab at idle, so subsequent check messages from
forgotten tabs are not handled properly.